+++ src/main/java/wordle/domain/BaseAnswerFormula.java
@@ -7,7 +7,7 @@ public class BaseAnswerFormula implements AnswerFormula {
 
     private final static LocalDate BASE = LocalDate.of(2021, 6, 19);
 
-    public long calculate(int wordCount) {
-        return ChronoUnit.DAYS.between(BASE, LocalDate.now()) % wordCount;
+    public int calculate(int wordCount) {
+        return (int) ChronoUnit.DAYS.between(BASE, LocalDate.now()) % wordCount;
     }
 }
